+++ CHANGELOG.md
@@ -155,6 +155,10 @@ All notable changes to this project will be documented in this file. This change
 ### 2023-08-15
 - Move telegram key to environment variable
 
+### 2023-09-19
+- Remove clj-time library
+- Change asset news database to use XTDB
+
 ## [0.1.1] - 2023-03-25
 ### Changed